+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 4.1.26.2 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Fixed bug in `DashboardCollection` stored `query`, instead list of `<dict>`, was sometimes list of `<instance>`
+ (???), added upgrade step to 4108, this is necessary for `plone.restapi` to serialize `DashboardCollection` to json
 
 4.1.26.1 (2020-06-12)
 ---------------------